feat: overhaul sintetizador-decomp v3.0.0

+++ app/services/handlers.py
@@ -1,45 +1,49 @@
 import pathlib
 import shutil
-from app.model.settings import Settings
+
 import app.domain.commands as commands
-from app.services.unitofwork import AbstractUnitOfWork
-from app.services.synthesis.system import SystemSynthetizer
+from app.model.settings import Settings
 from app.services.synthesis.execution import ExecutionSynthetizer
-from app.services.synthesis.scenarios import ScenarioSynthetizer
 from app.services.synthesis.operation import OperationSynthetizer
 from app.services.synthesis.policy import PolicySynthetizer
+from app.services.synthesis.scenarios import ScenarioSynthetizer
+from app.services.synthesis.system import SystemSynthetizer
+from app.services.unitofwork import AbstractUnitOfWork
 
 
 def synthetize_system(
     command: commands.SynthetizeSystem, uow: AbstractUnitOfWork
-):
+) -> None:
     SystemSynthetizer.synthetize(command.variables, uow)
 
 
 def synthetize_execution(
     command: commands.SynthetizeExecution, uow: AbstractUnitOfWork
-):
+) -> None:
     ExecutionSynthetizer.synthetize(command.variables, uow)
 
 
 def synthetize_scenario(
     command: commands.SynthetizeScenario, uow: AbstractUnitOfWork
-):
+) -> None:
     ScenarioSynthetizer.synthetize(command.variables, uow)
 
 
 def synthetize_operation(
     command: commands.SynthetizeOperation, uow: AbstractUnitOfWork
-):
+) -> None:
     OperationSynthetizer.synthetize(command.variables, uow)
 
 
 def synthetize_policy(
     command: commands.SynthetizePolicy, uow: AbstractUnitOfWork
-):
+) -> None:
     PolicySynthetizer.synthetize(command.variables, uow)
 
 
-def clean():
-    path = pathlib.Path(Settings().basedir).joinpath(Settings().synthesis_dir)
+def clean() -> None:
+    settings = Settings()
+    if settings.basedir is None:
+        return
+    path = pathlib.Path(settings.basedir).joinpath(settings.synthesis_dir)
     shutil.rmtree(path)
